Update: - Adding ability to block a user. Works out decently so far, though I think I should remove the post before it gets to the screen...

+++ app/src/main/java/com/example/gogieats/screens/map/MapScreen.kt
@@ -372,7 +372,7 @@ private fun SheetContent(
                         postUser = userViewModel.user)
                 }
             } else {
-                itemsIndexed(posts.value.sortedByDescending { it.distance.toDouble() }
+                itemsIndexed(posts.value.sortedByDescending { it.distance }
                     .reversed()) { index, restaurant ->
                     Spacer(Modifier.size(10.dp))
                     RestaurantCard(restaurant, location) {
